strawgate-es-mcp/pyproject.toml
[project]
name = "strawgate_es_mcp"
version = "0.5.0"
description = "Strawgate's Es Mcp project"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.2",
    "asyncclick>=8.1.8",
    "elasticsearch>=8.18.0,<9.0.0",
    "fastmcp>=2.3.5",
    "mcp>=1.9.0",
    "pytest-asyncio>=1.0.0",
    "python-dotenv>=1.1.0",
]

[tool.poetry]
packages = [
    { include = "strawgate_es_mcp", from = "src" }
]
[tool.hatch.metadata]
allow-direct-references=true

[project.scripts]
strawgate-es-mcp = "strawgate_es_mcp.main:run_mcp"


[dependency-groups]
dev = [
  "pytest",
  "pytest-mock",
  "ruff",
]
lint = [
  "ruff"
]

[build-system]
requires = ["hatchling", "uv-dynamic-versioning>=0.7.0"]
build-backend = "hatchling.build"

[tool.ruff]
extend="../pyproject.toml"

[tool.pyright]
pythonVersion = "3.13"
extraPaths = ["src/"]
include = ["src/"]
exclude = [
    "**/archive/**",
    "**/node_modules/**",
    "**/__pycache__/**",
    "**/.venv/**",
    ".venv",
    "**/.pytest_cache/**",
    "**/.ruff_cache/**",
    "**/uv/python/**",
]
reportMissingTypeStubs = false
reportExplicitAny = false

executionEnvironments = [
    { root = "src/tests", reportPrivateUsage = false},
]
